Update docs with snake_case attributes explanation
I ran in to this and it took me a while to work out what I needed to do to get snake_case attributes to work properly, especially with your test helpers. As it turned out, you already had it covered and it's very simple to do. I thought a little explanation might save others some time.  I have also created a PR on the demo package to include a snake_case attribute.